+++ darts/models/forecasting/regression_model.py
@@ -897,7 +897,7 @@ class RegressionModel(GlobalForecastingModel):
         future_covariates: Optional[Sequence[TimeSeries]] = None,
         num_samples: int = 1,
         start: Optional[Union[pd.Timestamp, float, int]] = None,
-        start_format: Literal["point", "index"] = "point",
+        start_format: Literal["positional_index", "value_index"] = "value_index",
         forecast_horizon: int = 1,
         stride: int = 1,
         overlap_end: bool = False,