+++ CHANGELOG.md
@@ -12,7 +12,7 @@ but cannot always guarantee backwards compatibility. Changes that may **break co
 
 **Improvement**
 - `TimeSeries` with a `RangeIndex` starting in the negative start are now supported by `historical_forecasts`. [#1866](https://github.com/unit8co/darts/pull/1866) by [Antoine Madrona](https://github.com/madtoinou).
-- Added a new argument `start_format` to `historical_forecasts`, `start` can now be provided as an absolute index (positive or negative) instead of a point of the time index. [#1866](https://github.com/unit8co/darts/pull/1866) by [Antoine Madrona](https://github.com/madtoinou).
+- 🔴 Added a new argument `start_format` to `historical_forecasts`, `start` can now be provided as a 'positional index' (positive or negative) or a point of the time index ('value index', default behavior). [#1866](https://github.com/unit8co/darts/pull/1866) by [Antoine Madrona](https://github.com/madtoinou).
 
 **Fixed**
 - Fixed a bug in `TimeSeries.from_dataframe()` when using a pandas.DataFrame with `df.columns.name != None`. [#1938](https://github.com/unit8co/darts/pull/1938) by [Antoine Madrona](https://github.com/madtoinou).